Move MapWithExpiration to its own file
