+++ packages/drivers/driver-base/src/index.ts
@@ -4,3 +4,4 @@
  */
 
 export { DocumentDeltaConnection } from "./documentDeltaConnection";
+export { MapWithExpiration } from "./mapWithExpiration";
